User histories 1-5
Commit with the user histories 1 to 5 completed

+++ js/initialization.js
@@ -1,25 +1,28 @@
-//alert('Welcome to Comics Armando!');
 
-// agregamos los datos al local storage
-localStorage.clear();
-console.log('limpiando local storage...');
+// we add data to the local storage
+console.log('Initializing local storage...');
 //rol
-var roles = [{'rolId':1,'rolName': 'admin','rolDesc': 'administrador del sitio'},
-			 {'rolId':2,'rolName': 'normal','rolDesc': 'usuario normal del sitio'}];
+var roles = [{'rolId':1,'rolName': 'administrator','rolDesc': 'website administrator'},
+			 {'rolId':2,'rolName': 'user','rolDesc': 'website user'}];
 localStorage.setItem('roles', JSON.stringify(roles));
-console.log('insercion roles...');
+console.log('adding roles...');
 
-//usuarios
-var usuarios = [{'userId':1,'usrName':'sheldon','usrPwd':'bazinga','usrEmail':'shledon@gmail.com','usrRol':'admin'},
-				{'userId':2,'usrName':'armando','usrPwd':'password','usrEmail':'armando@gmail.com','usrRol':'normal'}];
-localStorage.setItem('usuarios', JSON.stringify(usuarios));
-console.log('insercion usuarios...');
+//users
+var existUsrs= JSON.parse(localStorage.getItem('usuarios'));
+if(!existUsrs)
+{
+	var usuarios = [{'userId':1,'usrName':'sheldon','usrPwd':'bazinga','usrEmail':'shledon@gmail.com','usrRol':'administrator'},
+				{'userId':2,'usrName':'armando','usrPwd':'password','usrEmail':'armando@gmail.com','usrRol':'user'}];
+	localStorage.setItem('usuarios', JSON.stringify(usuarios));
+}
+
+console.log('adding users...');
 
 //roles
-var generos=[{'gnrId':1,'gnrName':'mutantes','gnrDesc':'Comics sobre mutantes'},
-			{'gnrId':2,'gnrName':'extraterrestres','gnrDesc':'Comics sobre extraterrestres'}];
+var generos=[{'gnrId':1,'gnrName':'mutantes','gnrDesc':'Comics about mutants'},
+			{'gnrId':2,'gnrName':'extraterrestres','gnrDesc':'Comics about aliens'}];
 localStorage.setItem('generos', JSON.stringify(generos));
-console.log('insercion generos...');
+console.log('adding geners...');
 
 //rate
 var rates=[{'rateId':1,'rateName':'5 estrellas','rateValue':5},
@@ -28,4 +31,4 @@ var rates=[{'rateId':1,'rateName':'5 estrellas','rateValue':5},
 			{'rateId':4,'rateName':'2 estrellas','rateValue':2},
 			{'rateId':5,'rateName':'1 estrellas','rateValue':1}];
 localStorage.setItem('rates', JSON.stringify(rates));
-console.log('insercion rates....');
+console.log('adding rates....');
